refactor: make iterator a non-default export and add replaceLongKeys to test utils
