+++ packages/article-flag/__tests__/article-flag-content.native.js
@@ -4,7 +4,7 @@ import {
   addSerializers,
   enzymeTreeSerializer
 } from "@times-components/jest-serializer";
-import iterator from "@times-components/test-utils";
+import { iterator } from "@times-components/test-utils";
 import ArticleFlag from "../src/article-flag";
 
 export default () => {
