+++ packages/article-flag/__tests__/web/article-flag-content.web.test.js
@@ -13,7 +13,7 @@ import {
   stylePrinter,
   rnwTransform
 } from "@times-components/jest-serializer";
-import iterator from "@times-components/test-utils";
+import { iterator } from "@times-components/test-utils";
 import ArticleFlag from "../../src/article-flag";
 
 addSerializers(
