+++ packages/test-utils/__tests__/android/iterator.test.js
@@ -5,7 +5,7 @@ import {
   minimalNativeTransform,
   print
 } from "@times-components/jest-serializer";
-import shared from "../shared";
+import shared from "../shared-iterator";
 
 addSerializers(
   expect,
