+++ packages/test-utils/package.json
@@ -33,7 +33,6 @@
   "homepage": "https://github.com/newsuk/times-components#readme",
   "devDependencies": {
     "@times-components/jest-configurator": "1.0.9",
-    "@times-components/jest-serializer": "1.0.9",
     "@times-components/webpack-configurator": "1.0.1",
     "babel-preset-react-native": "4.0.0",
     "enzyme": "3.3.0",
@@ -49,6 +48,9 @@
     "react-native": "0.54.2",
     "webpack": "4.6.0",
     "webpack-cli": "2.1.4"
+  },
+  "dependencies": {
+    "@times-components/jest-serializer": "1.0.9"
   },
   "publishConfig": {
     "access": "public"
