+++ packages/test-utils/src/index.js
@@ -1,8 +1,4 @@
-/* global it */
-const iterator = tests =>
-  tests.forEach(({ name, test }, indx) => {
-    const index = indx + 1;
-    it(`${index}. ${name.toLowerCase()}`, () => test());
-  });
+import iterator from "./iterator";
+import replaceLongKeys from "./replace-long-keys";
 
-export default iterator;
+export { iterator, replaceLongKeys };
