+++ packages/test-utils/src/replace-long-keys.js
@@ -1,0 +1,20 @@
+import { createHash } from "crypto";
+import { replacePropTransform } from "@times-components/jest-serializer";
+
+const hash = v =>
+  createHash("md5")
+    .update(v)
+    .digest("hex");
+
+// const longKeys = new Set(["d", "viewBox", "points"]);
+
+export default longKeysSet => {
+  if (longKeysSet instanceof Set) {
+    return replacePropTransform(
+      (value, key) =>
+        longKeysSet.has(key) ? hash(JSON.stringify(value)) : value
+    );
+  }
+
+  return null;
+};
