+++ packages/video-label/__tests__/video-label.native.js
@@ -8,7 +8,7 @@ import {
   minimalNativeTransform,
   print
 } from "@times-components/jest-serializer";
-import iterator from "@times-components/test-utils";
+import { iterator } from "@times-components/test-utils";
 import VideoLabel from "../src/video-label";
 
 export default () => {
