+++ packages/video-label/__tests__/web/video-label.web.test.js
@@ -12,7 +12,7 @@ import {
   replaceTransform,
   rnwTransform
 } from "@times-components/jest-serializer";
-import iterator from "@times-components/test-utils";
+import { iterator } from "@times-components/test-utils";
 import VideoLabel from "../../src/video-label";
 
 addSerializers(
