+++ packages/video/__tests__/web/video.web.test.js
@@ -6,7 +6,7 @@ import {
   minimaliseTransform,
   print
 } from "@times-components/jest-serializer";
-import iterator from "@times-components/test-utils";
+import { iterator } from "@times-components/test-utils";
 import Video from "../../src/video";
 import defaultVideoProps from "../default-video-props";
